chore: move electron utils file to utils/renderer and add to package.json
